refactor(layout): deepen pipeline entrypoints
