+++ lib/layout/collaboration/index.js
@@ -1,5 +1,4 @@
-import { createCollaborationLayoutContext } from './Context.js';
-import { runCollaborationPipeline } from './Pipeline.js';
+import { createLayout } from '../geometry/index.js';
 import { compactParticipantRows } from './steps/compactParticipantRows.js';
 import { layoutParticipants } from './steps/layoutParticipants.js';
 import { orderParticipants } from './steps/orderParticipants.js';
@@ -26,19 +25,44 @@ export const COLLABORATION_LAYOUT_STEPS = Object.freeze([
 ]);
 
 export function layoutCollaboration(collaboration, options = {}) {
-  const context = runCollaborationPipeline(
-    createCollaborationLayoutContext(collaboration, {
-      ...options,
-      steps: options.steps || COLLABORATION_LAYOUT_STEPS
-    })
-  );
+  const steps = options.steps || COLLABORATION_LAYOUT_STEPS;
+  const context = createCollaborationLayoutContext(collaboration, {
+    ...options,
+    steps
+  });
+  const completed = steps.reduce((current, runStep) => {
+    return runStep(current);
+  }, context);
 
   return {
-    layout: context.layout,
-    warnings: context.warnings
+    layout: completed.layout,
+    warnings: completed.warnings
   };
 }
 
-export {
-  createCollaborationLayoutContext
-} from './Context.js';
+function createCollaborationLayoutContext(
+    collaboration,
+    {
+      expandedIds = new Set(),
+      steps
+    }) {
+  return {
+    collaboration,
+    options: {
+      expandedIds,
+      steps
+    },
+    participants: {
+      layouts: new Map(),
+      anchorPositioned: new Set(),
+      expandable: new Set(),
+      order: []
+    },
+    routing: {
+      endpointDirections: new Map(),
+      channelOffsets: new Map()
+    },
+    layout: createLayout(collaboration),
+    warnings: []
+  };
+}
