+++ lib/layout/collaboration/steps/placeArtifacts.js
@@ -2,7 +2,7 @@ import { is } from '../../../di/DiUtil.js';
 import { placeArtifacts as layoutArtifacts } from '../../artifacts/index.js';
 import { isArtifact } from '../../bpmn/Predicates.js';
 import { layoutGroups } from '../../groups/LayoutGroups.js';
-import { createLayoutRecord } from '../../process/index.js';
+import { createLayoutRecord } from '../../process/LayoutRecord.js';
 
 /**
  * @typedef {import('../../Types.js').CollaborationLayoutContext} CollaborationLayoutContext
