--- lib/layout/process/Pipeline.js
@@ -1,13 +1,0 @@
-/**
- * @typedef {import('../Types.js').ProcessLayoutContext} ProcessLayoutContext
- */
-
-/**
- * @param {ProcessLayoutContext} context
- * @returns {ProcessLayoutContext}
- */
-export function runProcessPipeline(context) {
-  return context.options.steps.reduce((current, runStep) => {
-    return runStep(current);
-  }, context);
-}
