+++ lib/layout/process/index.js
@@ -1,5 +1,4 @@
-import { createProcessLayoutContext } from './Context.js';
-import { runProcessPipeline } from './Pipeline.js';
+import { createLayout } from '../geometry/index.js';
 import { analyzeSemantics } from './steps/analyzeSemantics.js';
 import { extractElements } from './steps/extractElements.js';
 import { layoutChildScopes } from './steps/layoutChildScopes.js';
@@ -34,19 +33,67 @@ export const PROCESS_LAYOUT_STEPS = Object.freeze([
  * Callers may provide an internally customized step list.
  */
 export function layoutProcessScope(scope, options = {}) {
-  const context = runProcessPipeline(createProcessLayoutContext(scope, {
+  const steps = options.steps || PROCESS_LAYOUT_STEPS;
+  const layoutScope = (childScope, childOptions = {}) => {
+    return layoutProcessScope(childScope, {
+      ...childOptions,
+      steps
+    });
+  };
+  const context = createProcessLayoutContext(scope, {
     ...options,
-    steps: options.steps || PROCESS_LAYOUT_STEPS
-  }));
+    layoutScope,
+    steps
+  });
+  const completed = steps.reduce((current, runStep) => {
+    return runStep(current);
+  }, context);
 
   return {
-    layout: context.layout,
-    warnings: context.warnings
+    layout: completed.layout,
+    warnings: completed.warnings
   };
 }
 
-export {
-  createLayoutRecord,
-  createProcessLayoutContext
-} from './Context.js';
 export { getParticipantContainerBounds } from './placement/ParticipantBounds.js';
+
+function createProcessLayoutContext(
+    scope,
+    {
+      expandedIds = new Set(),
+      participantProcess = false,
+      messageFlowEndpointDirections = new Map(),
+      steps,
+      layoutScope
+    }) {
+  return {
+    scope,
+    options: {
+      expandedIds,
+      participantProcess,
+      messageFlowEndpointDirections,
+      steps,
+      layoutScope
+    },
+    elements: {
+      groups: [],
+      sequenceFlows: [],
+      associations: []
+    },
+    graph: {
+      nodes: [],
+      edges: [],
+      boundaryEdges: []
+    },
+    semantics: {
+      policy: null,
+      ranks: null
+    },
+    placement: {
+      records: [],
+      recordsByElement: new Map()
+    },
+    layout: createLayout(scope),
+    warnings: []
+  };
+}
