+++ lib/layout/process/steps/extractElements.js
@@ -1,6 +1,6 @@
 import { is } from '../../../di/DiUtil.js';
 import { isArtifact } from '../../bpmn/Predicates.js';
-import { createLayoutRecord } from '../Context.js';
+import { createLayoutRecord } from '../LayoutRecord.js';
 
 /**
  * @typedef {import('../../Types.js').ProcessLayoutContext} ProcessLayoutContext
