+++ lib/layout/process/steps/layoutChildScopes.js
@@ -1,7 +1,5 @@
 import { is } from '../../../di/DiUtil.js';
 import { getExtents } from '../../geometry/index.js';
-import { createProcessLayoutContext } from '../Context.js';
-import { runProcessPipeline } from '../Pipeline.js';
 import { sizeExpandedSubProcess } from '../placement/ExpandedSubProcess.js';
 
 /**
@@ -17,8 +15,8 @@ export function layoutChildScopes(context) {
   const { records } = context.placement;
   const {
     expandedIds,
+    layoutScope,
     messageFlowEndpointDirections,
-    steps
   } = context.options;
 
   for (const record of records) {
@@ -26,17 +24,16 @@ export function layoutChildScopes(context) {
       continue;
     }
 
-    const childContext = runProcessPipeline(createProcessLayoutContext(
+    const childResult = layoutScope(
       record.element,
       {
         expandedIds,
-        messageFlowEndpointDirections,
-        steps
+        messageFlowEndpointDirections
       }
-    ));
+    );
 
-    record.child = childContext.layout;
-    warnings.push(...childContext.warnings);
+    record.child = childResult.layout;
+    warnings.push(...childResult.warnings);
 
     if (record.expanded) {
       record.size = sizeExpandedSubProcess(getExtents(record.child));
